refactor(migration): one home for the primitives every entry point needs
`dnsEncodeName` and `errorMessageChain` each existed twice, once in the
CLI and once beside the code that needed them. They are the smallest of
the shared pieces, and moving them first gives the rest of the tooling a
module to import from that reaches back into nothing.
